modify python files for routing system
